requirements.txt
Mastodon.py
pathlib
wordfilter
